Revamp imports by using several aliases
To better keep the implementation separate from the declaration, use several import aliases in TS to allow consumers to import "@/{items}" instead of requiring specific files
