+++ src/app/keyboardShortcuts.ts
@@ -1,5 +1,5 @@
 import hotkeys from 'hotkeys-js';
-import { appRoutedCommands, RoutedCommand } from '../framework/appRoutedCommands';
+import { appRoutedCommands, RoutedCommand } from '@/framework/appRoutedCommands';
 
 function trigger(shortcut, command: RoutedCommand) {
   hotkeys(shortcut, (evt) => {
