+++ src/components/design/control-editor.e.tsx
@@ -1,13 +1,13 @@
 import { h } from 'preact';
-import { determineEditStyle, calculateSnapTo } from '../../framework/positioner';
-import { Anchor, Point, AnchoredBoundary, IStoredPositionInfo } from '../../framework/layout';
-import { registerUndoHandler } from '../../framework/undoRedo';
+import { determineEditStyle, calculateSnapTo } from '@/framework/positioner';
+import { Anchor, Point, AnchoredBoundary, IStoredPositionInfo } from '@/framework/layout';
+import { registerUndoHandler } from '@/framework/undoRedo';
 import { ControlContainer } from './control-container.e';
 import { DragHandle } from './drag-handle.e';
 import { DesignSurfaceElement } from './design-surface.e';
 
 import './control-editor.css';
-import { UniqueId } from '../../framework/util';
+import { UniqueId } from '@/framework/util';
 import { CustomHtmlJsxElement, customElement } from '@friendly/elements/CustomHtmlJsxElement';
 
 @customElement(ControlEditor.tagName)
