+++ src/components/design/drag-handle.e.tsx
@@ -1,4 +1,4 @@
-import { Anchor } from '../../framework/layout';
+import { Anchor } from '@/framework/layout';
 import { CustomHtmlElement, customElement, property } from '@friendly/elements/CustomHtmlElement';
 
 @customElement(DragHandle.tagName)
@@ -30,11 +30,11 @@ export class DragHandle extends CustomHtmlElement {
       background-color: rgba(0, 0, 0, 0.5);
       z-index: 10010;
       position: absolute;
-    
+
       --handle-offset: -3px;
       --handle-size: 6px;
     }
-    
+
     .drag-handle-nw,
     .drag-handle-ne,
     .drag-handle-se,
@@ -42,75 +42,75 @@ export class DragHandle extends CustomHtmlElement {
       width: var(--handle-size);
       height: var(--handle-size);
     }
-    
+
     .drag-handle-nw {
       left: var(--handle-offset);
       top: var(--handle-offset);
-    
+
       cursor: nw-resize;
     }
-    
+
     .drag-handle-ne {
       right: var(--handle-offset);
       top: var(--handle-offset);
-    
+
       cursor: ne-resize;
     }
-    
+
     .drag-handle-se {
       right: var(--handle-offset);
       bottom: var(--handle-offset);
-    
+
       cursor: se-resize;
     }
-    
+
     .drag-handle-sw {
       left: var(--handle-offset);
       bottom: var(--handle-offset);
-    
+
       cursor: sw-resize;
     }
-    
+
     .drag-handle-east,
     .drag-handle-west,
     .drag-handle-north,
     .drag-handle-south {
       background: rgba(0, 0, 0, 0);
     }
-    
+
     .drag-handle-north {
       right: var(--handle-offset);
       top: var(--handle-offset);
       left: var(--handle-offset);
       height: var(--handle-size);
-    
+
       cursor: n-resize;
     }
-    
+
     .drag-handle-east {
       right: var(--handle-offset);
       top: var(--handle-offset);
       bottom: var(--handle-offset);
       width: var(--handle-size);
-    
+
       cursor: e-resize;
     }
-    
+
     .drag-handle-west {
       left: var(--handle-offset);
       top: var(--handle-offset);
       bottom: var(--handle-offset);
       width: var(--handle-size);
-    
+
       cursor: w-resize;
     }
-    
+
     .drag-handle-south {
       right: var(--handle-offset);
       bottom: var(--handle-offset);
       left: var(--handle-offset);
       height: var(--handle-size);
-    
+
       cursor: s-resize;
     }
     `;
