+++ src/controls/core/@control.tsx
@@ -1,0 +1,9 @@
+export * from './controlRegistry';
+export * from './controlProperties';
+export * from './propertyBag';
+export * from './Control';
+
+/**
+ * Module to be imported when implementing new controls or interacting with controls or their corresponding apis,
+ * descriptors, or properties.
+ */
