+++ src/controls/core/controlRegistry.ts
@@ -1,13 +1,8 @@
-import {
-  addValue,
-  Control,
-  getControlPropertiesFor,
-  IControlProperty,
-  ISerializedPropertyBag,
-  tryGetValue,
-} from 'src/controls/Control';
-import { IStoredPositionInfo } from '../framework/layout';
-import { LocalizedString } from '../framework/localization';
+import { IStoredPositionInfo } from '@/framework/layout';
+import { LocalizedString } from '@/framework/localization';
+import { Control } from './Control';
+import { getControlPropertiesFor, IControlProperty } from './controlProperties';
+import { addValue, ISerializedPropertyBag, tryGetValue } from './propertyBag';
 
 /**
  * Holds information about the controls that can be edited via the design surface.  It is
