+++ src/controls/core/propertyUndoRedo.tsx
@@ -1,9 +1,9 @@
-import { registerUndoHandler } from '../../framework/undoRedo';
-import { UniqueId } from '../../framework/util';
-import { getFocusCount, FocusCount } from '../../framework/focusService';
+import { registerUndoHandler } from '@/framework/undoRedo';
+import { UniqueId } from '@/framework/util';
+import { getFocusCount, FocusCount } from '@/framework/focusService';
 import { ControlContainer } from 'src/components/design/control-container.e';
-import { RoutedEventDescriptor } from 'src/framework/routedEvents';
-import { IControlProperty } from '../controlProperties';
+import { RoutedEventDescriptor } from '@/framework/routedEvents';
+import { IControlProperty } from '@/control';
 
 interface SetPropertyUndoArgs {
   id: UniqueId;
