+++ src/controls/~RootControl.tsx
@@ -1,7 +1,12 @@
-import { Control, IControlSerializedData, implementProperty } from './Control';
-import { IControlDescriptor, ReflectionBasedDescriptor } from './controlRegistry';
-import { UniqueId } from '../framework/util';
-import { BackgroundColorProperty } from './properties/@commonProperties';
+import {
+  Control,
+  IControlDescriptor,
+  IControlSerializedData,
+  implementProperty,
+  ReflectionBasedDescriptor,
+} from '@/control';
+import { UniqueId } from '@/framework/util';
+import { BackgroundColorProperty } from '@/control/standardProperties';
 
 export class RootControl extends Control {
   private containerElement: HTMLDivElement;
